Extract configuration and rejection filter modules
